+++ desafioMoviesTMDB/desafioMoviesTMDB/Views/MainView/Components/AboutTheMovieView/Components/HeartButtonView.swift
@@ -24,15 +24,3 @@ struct HeartButtonView: View {
         }
     }
 }
-
-struct HeartButtonView_Previews: PreviewProvider {
-    
-    static var previews: some View {
-        
-        ZStack {
-            Color.black
-            HeartButtonView()
-        }
-        
-    }
-}
