+++ desafioMoviesTMDB/desafioMoviesTMDB/Views/MainView/Components/ProgressViewScreen.swift
@@ -1,0 +1,22 @@
+//
+//  ProgressViewScreen.swift
+//  desafioMoviesTMDB
+//
+//  Created by Diego Llopis on 11/11/21.
+//
+
+import SwiftUI
+
+struct ProgressViewScreen: View {
+    
+    var body: some View {
+        
+        ZStack {
+            Color.black
+            
+            ProgressView()
+                .progressViewStyle(CircularProgressViewStyle(tint: .red))
+                .scaleEffect(2)
+        }
+    }
+}
